[move-prover] Transferred strong edge borrow analysis out of experimental pipeline
