+++ language/move-prover/bytecode/src/lib.rs
@@ -10,7 +10,6 @@ pub mod access_path;
 pub mod access_path_trie;
 pub mod annotations;
 pub mod borrow_analysis;
-pub mod borrow_analysis_v2;
 pub mod clean_and_optimize;
 pub mod compositional_analysis;
 pub mod data_invariant_instrumentation;
@@ -26,7 +25,6 @@ pub mod graph;
 pub mod livevar_analysis;
 pub mod loop_analysis;
 pub mod memory_instrumentation;
-pub mod memory_instrumentation_v2;
 pub mod mut_ref_instrumentation;
 pub mod options;
 pub mod packed_types_analysis;
