+++ language/move-prover/bytecode/tests/testsuite.rs
@@ -64,7 +64,7 @@ fn get_tested_transformation_pipeline(
             pipeline.add_processor(MutRefInstrumenter::new());
             pipeline.add_processor(ReachingDefProcessor::new());
             pipeline.add_processor(LiveVarAnalysisProcessor::new());
-            pipeline.add_processor(BorrowAnalysisProcessor::new());
+            pipeline.add_processor(BorrowAnalysisProcessor::new(false));
             Ok(Some(pipeline))
         }
         "memory_instr" => {
@@ -73,7 +73,7 @@ fn get_tested_transformation_pipeline(
             pipeline.add_processor(MutRefInstrumenter::new());
             pipeline.add_processor(ReachingDefProcessor::new());
             pipeline.add_processor(LiveVarAnalysisProcessor::new());
-            pipeline.add_processor(BorrowAnalysisProcessor::new());
+            pipeline.add_processor(BorrowAnalysisProcessor::new(false));
             pipeline.add_processor(MemoryInstrumentationProcessor::new());
             Ok(Some(pipeline))
         }
@@ -83,7 +83,7 @@ fn get_tested_transformation_pipeline(
             pipeline.add_processor(MutRefInstrumenter::new());
             pipeline.add_processor(ReachingDefProcessor::new());
             pipeline.add_processor(LiveVarAnalysisProcessor::new());
-            pipeline.add_processor(BorrowAnalysisProcessor::new());
+            pipeline.add_processor(BorrowAnalysisProcessor::new(false));
             pipeline.add_processor(MemoryInstrumentationProcessor::new());
             pipeline.add_processor(CleanAndOptimizeProcessor::new());
             Ok(Some(pipeline))
@@ -94,7 +94,7 @@ fn get_tested_transformation_pipeline(
             pipeline.add_processor(MutRefInstrumenter::new());
             pipeline.add_processor(ReachingDefProcessor::new());
             pipeline.add_processor(LiveVarAnalysisProcessor::new());
-            pipeline.add_processor(BorrowAnalysisProcessor::new());
+            pipeline.add_processor(BorrowAnalysisProcessor::new(false));
             pipeline.add_processor(MemoryInstrumentationProcessor::new());
             pipeline.add_processor(CleanAndOptimizeProcessor::new());
             pipeline.add_processor(UsageProcessor::new());
@@ -108,7 +108,7 @@ fn get_tested_transformation_pipeline(
             pipeline.add_processor(MutRefInstrumenter::new());
             pipeline.add_processor(ReachingDefProcessor::new());
             pipeline.add_processor(LiveVarAnalysisProcessor::new());
-            pipeline.add_processor(BorrowAnalysisProcessor::new());
+            pipeline.add_processor(BorrowAnalysisProcessor::new(false));
             pipeline.add_processor(MemoryInstrumentationProcessor::new());
             pipeline.add_processor(CleanAndOptimizeProcessor::new());
             pipeline.add_processor(UsageProcessor::new());
@@ -124,7 +124,7 @@ fn get_tested_transformation_pipeline(
             pipeline.add_processor(MutRefInstrumenter::new());
             pipeline.add_processor(ReachingDefProcessor::new());
             pipeline.add_processor(LiveVarAnalysisProcessor::new());
-            pipeline.add_processor(BorrowAnalysisProcessor::new());
+            pipeline.add_processor(BorrowAnalysisProcessor::new(false));
             pipeline.add_processor(MemoryInstrumentationProcessor::new());
             pipeline.add_processor(CleanAndOptimizeProcessor::new());
             pipeline.add_processor(UsageProcessor::new());
